Docker doesn't start if you dont have the binary

+++ tools/Test-CentOS7/run_test_in_docker.sh
@@ -1,5 +1,13 @@
 #!/bin/bash
-cp ../../pgglaskugel .
+if [ -f ../../pgglaskugel ]
+  then
+    cp ../../pgglaskugel .
+  else
+    echo "ERROR: Could not find pgglaskugel"
+    exit 1
+fi
+
+echo "GPG takes longer when generating keys if you don't have enough entropy"
 sudo docker build -t=pgglaskugelcentos7 .
 sudo docker run -it pgglaskugelcentos7
 rm pgglaskugel
